examples/model_hosting_1/requirements.txt
Django
protobuf
requests
loguru
tensorflow-serving-api
cryptography~=2.8
django-cors-headers
drf-jwt
pyjwt

